Former-commit-id: 29b5e8964a1653fe5cb769375dbcbb3ec8bb7bcf

+++ cocos2d-ui-tests/tests/CCResponderTest.m
@@ -45,7 +45,7 @@
 
 - (void) update:(CCTime)delta
 {
-    _lblNumTouches.string = [NSString stringWithFormat:@"Num touches: %d", _currentTouches.count];
+    _lblNumTouches.string = [NSString stringWithFormat:@"Num touches: %d", (unsigned int)_currentTouches.count];
 }
 
 - (void) touchBegan:(UITouch *)touch withEvent:(UIEvent *)event
